domain layer refactoring 완료 - useCase -> extensions로 명명 수정
