+++ adapter/dao/src/main/kotlin/team/bakkas/dao/repository/dynamo/ShopDynamoRepositoryImpl.kt
@@ -8,9 +8,8 @@ import reactor.core.publisher.Mono
 import software.amazon.awssdk.enhanced.dynamodb.DynamoDbAsyncTable
 import software.amazon.awssdk.enhanced.dynamodb.DynamoDbEnhancedAsyncClient
 import software.amazon.awssdk.enhanced.dynamodb.Key
-import software.amazon.awssdk.enhanced.dynamodb.TableSchema
 import team.bakkas.dynamo.shop.Shop
-import team.bakkas.dynamo.shop.usecases.softDelete
+import team.bakkas.dynamo.shop.extensions.softDelete
 import team.bakkas.repository.ifs.dynamo.ShopDynamoRepository
 
 /**
