+++ domain/dynamo/src/main/kotlin/team/bakkas/dynamo/reviewReply/extentions/ReviewReplyExtensions.kt
@@ -1,3 +1,3 @@
-package team.bakkas.dynamo.reviewReply.usecases
+package team.bakkas.dynamo.reviewReply.extentions
 
 // ReviewReply에 대한 usecase를 정의하는 코틀린 파일
